+++ src/main/scala/com/guizmaii/distances/DistanceApi.scala
@@ -7,6 +7,7 @@ import cats.temp.par.Par
 import com.guizmaii.distances.Types._
 import com.guizmaii.distances.providers.{CacheProvider, DistanceProvider, InMemoryCacheProvider}
 
+import scala.collection.breakOut
 import scala.concurrent.duration._
 
 class DistanceApi[AIO[_]: Par](distanceProvider: DistanceProvider[AIO], cacheProvider: CacheProvider[AIO])(implicit AIO: Async[AIO]) {
@@ -42,11 +43,11 @@ class DistanceApi[AIO[_]: Par](distanceProvider: DistanceProvider[AIO], cachePro
     else
       (geocoder.geocodePostalCode(origin), geocoder.geocodePostalCode(destination)).parMapN { case (o, d) => distance(o, d, travelModes) }.flatten
 
-  final def distances(paths: List[DirectedPath]): AIO[Map[(TravelMode, LatLong, LatLong), Distance]] = {
+  final def distances(paths: Seq[DirectedPath]): AIO[Map[(TravelMode, LatLong, LatLong), Distance]] = {
     val combinedDirectedPaths: List[DirectedPath] =
       paths
         .filter(_.travelModes.nonEmpty)
-        .combineDuplicatesOn { case DirectedPath(origin, destination, _) => (origin, destination) }(directedPathSemiGroup)
+        .combineDuplicatesOn { case DirectedPath(origin, destination, _) => (origin, destination) }(directedPathSemiGroup, breakOut)
 
     // TODO: Replace by the syntax extension when this issue is closed: https://github.com/typelevel/cats/issues/2255
     Parallel
@@ -74,7 +75,7 @@ object DistanceApi {
   final def apply[AIO[_]: Async: Par](provider: DistanceProvider[AIO], cacheProvider: CacheProvider[AIO]): DistanceApi[AIO] =
     new DistanceApi(provider, cacheProvider)
 
-  private[DistanceApi] final val directedPathSemiGroup: Semigroup[DirectedPath] = new Semigroup[DirectedPath] {
+  private[DistanceApi] implicit final val directedPathSemiGroup: Semigroup[DirectedPath] = new Semigroup[DirectedPath] {
     override def combine(x: DirectedPath, y: DirectedPath): DirectedPath =
       DirectedPath(origin = x.origin, destination = x.destination, (x.travelModes ++ y.travelModes).distinct)
   }